rust: do not panic on next write if chunk closure fails (#1441)
### Changelog
- rust: removed the condition for a panic if a subsequent write failed.
### Docs

None.
<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

Previously, if we attempted to close a chunk and this failed, the writer
would be left in an invalid state, with `self.writer = None`. This is no
longer the case. Instead we return an explicit error value for cases
where we can no longer write because the underlying file is invalid in
an (unknown) state.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->