Clarify schema and message encodings (#1473)
### Changelog

- Clarified definitions of `ros1msg`, `ros2msg` and `ros2idl` schema
encoding and `cdr` message encoding.

### Docs
None

### Description

The `cdr` message encoding linked to the wrong document and didn't
actually explain that it was specifically XCDR1/XCDR2 that it contained.

The `ros1msg` schema encoding (and by extension `ros2msg`) left open a
possible ambiguity with multiple messages sharing the same combination
of package and message name. This is now forbidden.

The `ros2idl` schema encoding did not specify how it was to be
interpreted. I have not been able to find any official ROS2
documentation about this, so I have added what I believe to be the most
reasonable interpretation.

Fixes: FIRE-97
Fixes: FIRE-180