rust: fix int underflow when chunk wasn't closed (#1480)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- rust: fix int underflow on chunk with invalid size

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

When a chunk doesn't close correctly (loss of power, crash, etc.) the
size that was written will be `u64::MAX`. If you try and read a file
like this with the Rust reader it breaks because the size of the chunk
is expected to be valid. In debug mode this will panic due to underflow,
in release mode this value will wrap and break something else down the
track.

This change explicitly uses checked calculations for the chunk size and
returns an `UnexpectedEoc` error if any of the calculations overflow.

Fixes #1479.
Fixes FIRE-202.