go: fix cli recover command error messages (#1483)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- Fixes the mcap recover command printing to stdout
- Fixes incorrect error message

### Docs

None

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

Currently the MCAP CLI prints error messages to stdout even when it is
being used to write the "recovered" file. This leads to the file in some
cases having error messages interleaved throughout it, which creates an
invalid this.

This is only an issue when redirecting stdout to a file, not when the
`--output` flag is used.