cli: fix support for piping stdin to cat
The cat command supports accepting input from stdin, which is useful for
building shell pipelines out of mcap data streams when combined with the
JSON transcoding support:

    cat demo.mcap | mcap cat --json | jq ...

Commit 431d1dce2 caused this behavior to break, because it caused a
default of useIndex=true to get inherited in the cat options assessment,
when "useIndex=false" was requested (as triggered by the stdin
handling). That commit seems to assume that the reader will skip the
index by default, which is not true at least today (and presumably not
then either).

With this new commit, correct behavior is restored:

    [~] $ cat demo.mcap | mcap cat | head -n 1
    1490149580103843113 /diagnostics [diagnostic_msgs/DiagnosticArray] [42 10 0 0 204 224 209 88 99 250]...

this patch also adds a test that confirms that the intent of the
original bad commit is still preserved.